// src/models/IPenalty.ts
export interface Penalty {
    id: string;
    mentor: Mentor;
    reason: string;
    date: number;
    session: Session;
    amountBefore: number;
}

export interface Mentor {
    id: string;
    name: string;
    status: string;
    degree: string;
    major: string;
    countryOfOrigin: string;
    languages: string[];
    campusInfo: CampusInfo;
    amountAvailable: number;
    totalWithdrawn: number;
    preferredCurrency: string;
    schedule: Schedule;
    disabled: boolean;
    countryOfStudy: string;
    lastWithdrawn: number;
    profileState: string;
}

export interface CampusInfo {
    verificationDocUrl: string;
    uniEmail: UniEmail;
    uniName: string;
    scores: Scores;
    specialisation: string;
    campusExperience: string;
    scholarship: string;
    placeOfStay: string;
    previousInstitute: string;
    courseName: string;
}

export interface UniEmail {
    email: string;
    verified: boolean;
}

export interface Scores {
    gpa: string;
    gmat: string;
    sat: string;
}

export interface Schedule {
    id: string;
    days: Days;
}
// ...
export interface Session {
    payment: Payment;
    discussionPoints: string;
    id: string;
    zoomLink: string;
    mentor: Mentor3;
    user: User2;
    date: number;
    slot: Slot;
}

export interface Payment {
    id: string;
    razorpayId: string;
    sessions: string[];
    currency: string;
    date: number;
    mentor: Mentor2;
    user: User;
    amount: number;
}
// ...
export interface Mentor3 {
    institute: string;
    campusPreference: string;
    profilePic: string;
    name: string;
    about: string;
    facebook: string;
    linkedIn: string;
    email: string;
    id: string;
    type: string;
    interests: string[];
    mentorMeta: MentorMeta3;
}
// ...
export interface User2 {
    institute: string;
    campusPreference: string;
    profilePic: string;
    name: string;
    about: string;
    facebook: string;
    linkedIn: string;
    email: string;
    id: string;
    type: string;
    interests: string[];
    mentorMeta: MentorMeta4;
}
// ...
export interface Slot {
    start: string;
    end: string;
    lastBookedFor: number;
}